Switch to gem-release from jeweler for gem version bumping in a manner compatible with an FBGraph::VERSION constant. Proper spec environment setup (all specs passing). Switch from rcov to simplecov (Ruby 1.9 compatibility).
