+++ Gemfile
@@ -1,20 +1,3 @@
 source :rubygems
 
-gem "activesupport"
-gem "hashie", ">= 1.0.0"
-gem "oauth2", ">= 0.5.0"
-gem "faraday", ">= 0.7.5"
-gem "json", ">= 1.0.0"
-gem "rest-client", :require => "rest_client"
-gem "i18n"
-
-group :development do
-  gem "jeweler", "~> 1.6.0"
-  gem "rake", "~> 0.9.2"
-  gem "bundler", "~> 1.0.0"
-  gem "fakeweb", "~> 1.3.0"
-  gem "rspec", ">= 2.6.0"
-  gem "rcov", ">= 0"
-  gem "rdoc", ">= 3.9.0"
-  gem "redgreen", "~> 1.2.2"
-end
+gemspec
