+++ Rakefile
@@ -1,32 +1,12 @@
-require 'rubygems'
-require 'bundler'
-begin
-  Bundler.setup(:default, :development)
-rescue Bundler::BundlerError => e
-  $stderr.puts e.message
-  $stderr.puts "Run `bundle install` to install missing gems"
-  exit e.status_code
-end
+# encoding: utf-8
+#!/usr/bin/env rake
+require "bundler/gem_tasks"
 require 'rake'
-
-require 'jeweler'
-Jeweler::Tasks.new do |gem|
-  gem.name = "fbgraph"
-  gem.homepage = "http://github.com/nsanta/fbgraph"
-  gem.license = "MIT"
-  gem.description = "A Gem for Facebook Open Graph API"
-  gem.summary = "A Gem for Facebook Open Graph API"
-  gem.author = "Nicolas Santa"
-  gem.email = "nicolas55ar@gmail.com"
-
-  # Dependencies in Gemfile.
-end
-Jeweler::RubygemsDotOrgTasks.new
 
 require 'rspec/core'
 require 'rspec/core/rake_task'
 RSpec::Core::RakeTask.new(:spec) do |spec|
-  spec.pattern = FileList['specs/**/*_spec.rb']
+  spec.pattern = FileList['spec/**/*_spec.rb']
 end
 
 RSpec::Core::RakeTask.new(:rcov) do |spec|
@@ -38,10 +18,12 @@ task :default => :spec
 
 require 'rdoc/task'
 Rake::RDocTask.new do |rdoc|
-  version = File.exist?('VERSION') ? File.read('VERSION') : ""
-
+  require File.expand_path('../lib/fbgraph/version', __FILE__)
   rdoc.rdoc_dir = 'rdoc'
-  rdoc.title = "anki-importer #{version}"
+  rdoc.title    = "SanitizeEmail #{FBGraph::VERSION}"
+  rdoc.options << '--line-numbers'
   rdoc.rdoc_files.include('README*')
   rdoc.rdoc_files.include('lib/**/*.rb')
 end
+
+Bundler::GemHelper.install_tasks
