+++ fbgraph.gemspec
@@ -1,11 +1,9 @@
-# Generated by jeweler
-# DO NOT EDIT THIS FILE DIRECTLY
-# Instead, edit Jeweler::Tasks in Rakefile, and run 'rake gemspec'
 # -*- encoding: utf-8 -*-
+require File.expand_path('../lib/fbgraph/version', __FILE__)
 
 Gem::Specification.new do |s|
   s.name = "fbgraph"
-  s.version = "1.10.0"
+  s.version = FBGraph::VERSION
 
   s.required_rubygems_version = Gem::Requirement.new(">= 0") if s.respond_to? :required_rubygems_version=
   s.authors = ["Nicolas Santa"]
@@ -13,8 +11,14 @@ Gem::Specification.new do |s|
   s.description = "A Gem for Facebook Open Graph API"
   s.email = "nicolas55ar@gmail.com"
   s.extra_rdoc_files = [
-    "README",
     "README.textile"
+  ]
+  s.files         = `git ls-files`.split($\)
+  s.executables   = s.files.grep(%r{^bin/}).map{ |f| File.basename(f) }
+  s.test_files    = s.files.grep(%r{^(test|spec|features)/})
+
+  s.extra_rdoc_files = [
+    "README",
   ]
   s.files = [
     ".bundle/config",
@@ -24,7 +28,6 @@ Gem::Specification.new do |s|
     "README.textile",
     "Rakefile",
     "TODO.list",
-    "VERSION",
     "fbgraph.gemspec",
     "lib/fbgraph.rb",
     "lib/fbgraph/authorization.rb",
@@ -65,12 +68,11 @@ Gem::Specification.new do |s|
       s.add_runtime_dependency(%q<json>, [">= 1.0.0"])
       s.add_runtime_dependency(%q<rest-client>, [">= 0"])
       s.add_runtime_dependency(%q<i18n>, [">= 0"])
-      s.add_development_dependency(%q<jeweler>, ["~> 1.6.0"])
       s.add_development_dependency(%q<rake>, ["~> 0.9.2"])
-      s.add_development_dependency(%q<bundler>, ["~> 1.0.0"])
+      s.add_development_dependency(%q<bundler>, ["> 1.0.0"])
       s.add_development_dependency(%q<fakeweb>, ["~> 1.3.0"])
       s.add_development_dependency(%q<rspec>, [">= 2.6.0"])
-      s.add_development_dependency(%q<rcov>, [">= 0"])
+      s.add_development_dependency(%q<simplecov>, [">= 0"])
       s.add_development_dependency(%q<rdoc>, [">= 3.9.0"])
       s.add_development_dependency(%q<redgreen>, ["~> 1.2.2"])
     else
@@ -81,12 +83,11 @@ Gem::Specification.new do |s|
       s.add_dependency(%q<json>, [">= 1.0.0"])
       s.add_dependency(%q<rest-client>, [">= 0"])
       s.add_dependency(%q<i18n>, [">= 0"])
-      s.add_dependency(%q<jeweler>, ["~> 1.6.0"])
       s.add_dependency(%q<rake>, ["~> 0.9.2"])
-      s.add_dependency(%q<bundler>, ["~> 1.0.0"])
+      s.add_dependency(%q<bundler>, ["> 1.0.0"])
       s.add_dependency(%q<fakeweb>, ["~> 1.3.0"])
       s.add_dependency(%q<rspec>, [">= 2.6.0"])
-      s.add_dependency(%q<rcov>, [">= 0"])
+      s.add_dependency(%q<simplecov>, [">= 0"])
       s.add_dependency(%q<rdoc>, [">= 3.9.0"])
       s.add_dependency(%q<redgreen>, ["~> 1.2.2"])
     end
@@ -98,12 +99,11 @@ Gem::Specification.new do |s|
     s.add_dependency(%q<json>, [">= 1.0.0"])
     s.add_dependency(%q<rest-client>, [">= 0"])
     s.add_dependency(%q<i18n>, [">= 0"])
-    s.add_dependency(%q<jeweler>, ["~> 1.6.0"])
     s.add_dependency(%q<rake>, ["~> 0.9.2"])
-    s.add_dependency(%q<bundler>, ["~> 1.0.0"])
+    s.add_dependency(%q<bundler>, ["> 1.0.0"])
     s.add_dependency(%q<fakeweb>, ["~> 1.3.0"])
     s.add_dependency(%q<rspec>, [">= 2.6.0"])
-    s.add_dependency(%q<rcov>, [">= 0"])
+    s.add_dependency(%q<simplecov>, [">= 0"])
     s.add_dependency(%q<rdoc>, [">= 3.9.0"])
     s.add_dependency(%q<redgreen>, ["~> 1.2.2"])
   end
