+++ lib/fbgraph/version.rb
@@ -1,0 +1,3 @@
+module FBGraph
+  VERSION = '1.10.0'
+end
