+++ spec/lib/fbauth/authorization_spec.rb
@@ -1,4 +1,4 @@
-require File.expand_path('../../../spec_helper', __FILE__)
+require "spec_helper"
 
 describe FBGraph do
   describe FBGraph::Authorization do
